@@ -3,7 +3,5 @@ class WelcomeController < ApplicationController
   def index
     render :layout => "welcome"
   end
-  def actions
-    render :layout => nil
-  end
+
 end
